Add solution for problem `D - Round Subset` Link: https://codeforces.com/contest/837/problem/D Tags: dp, math, *2100 Ref: https://codeforces.com/contest/837/submission/29260794

+++ codeforces/837/D.cpp
@@ -3,14 +3,14 @@ using namespace std;
 #define popCnt(x) (__builtin_popcountll(x))
 typedef long long Long;
 
-const short N = 201;
+const int N = 201;
 const Long MX = 1e18 + 1;
-const short LogMX = log(MX) / log(5) + 3;
-const short M = 2 * N * LogMX + 5;
+const int LogMX = log(MX) / log(5) + 3;
+const int M = 2 * N * LogMX + 5;
 
-short dp[2][N][M];
+int dp[2][N][M];
 
-short cnt5[N], cnt2[N];
+int cnt5[N], cnt2[N];
 
 int main() {
   ios_base::sync_with_stdio(0), cin.tie(0), cout.tie(0);
@@ -40,16 +40,16 @@ int main() {
     bool curr = i & 1;
     for (int rem = 1; rem < N; ++rem) {
       for (int c = 0; c < M; ++c) {
-        short leave = dp[!curr][rem][c];
+        int leave = dp[!curr][rem][c];
 
-        short new_cnt5 = max((short)0, short(c - M / 2)) + cnt5[i];
-        short new_cnt2 = max((short)0, short(M / 2 - c)) + cnt2[i];
+        int new_cnt5 = max(0, c - M / 2) + cnt5[i];
+        int new_cnt2 = max(0, M / 2 - c) + cnt2[i];
 
-        short new_c = M / 2 + new_cnt5 - new_cnt2;
-        new_c = max(new_c , short(0));
-        new_c = min(new_c , short(M - 1));
+        int new_c = M / 2 + new_cnt5 - new_cnt2;
+        new_c = max(new_c , 0);
+        new_c = min(new_c , M - 1);
         
-        short take = min(new_cnt5, new_cnt2) + dp[!curr][rem - 1][new_c];
+        int take = min(new_cnt5, new_cnt2) + dp[!curr][rem - 1][new_c];
 
         dp[curr][rem][c] = max(leave, take);
       }
